auth/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==0.19.0
passlib[bcrypt]
pymongo==3.12.0
pyjwt
python-multipart